Move install-related methods and classes to the ContentRepository project.
